[frontend] make defaultValues fields from entitySettings reusable
